flatten optimistic executioner into auctioneer
The "optimistic executioner" task was an extra
indirection that provided no clear benefit over
just running all stream directly in the auctioneer
task.
