+++ crates/astria-auctioneer/src/sequencer_channel.rs
@@ -28,12 +28,17 @@ use tonic::transport::Channel;
 
 use crate::block::Commitment;
 
+pub(crate) fn open(endpoint: &str) -> eyre::Result<SequencerChannel> {
+    SequencerChannel::create(&endpoint)
+        .wrap_err_with(|| format!("failed to create a gRPC channel to Sequencer at `{endpoint}`"))
+}
+
 pub(crate) struct SequencerChannel {
     inner: Channel,
 }
 
 impl SequencerChannel {
-    pub(crate) fn create(uri: &str) -> eyre::Result<Self> {
+    fn create(uri: &str) -> eyre::Result<Self> {
         let channel = Channel::from_shared(uri.to_string())
             .wrap_err("failed to open a channel to the provided uri")?
             .connect_timeout(Duration::from_secs(5))
